Added Shields
Added the abstract and concrete classes for the shields

+++ Еscape from the labyrinth of death/Еscape from the labyrinth of death/Interfaces/IItem.cs
@@ -10,10 +10,10 @@
     }
     interface IItem
     {
-        int BonusToHealth { get; }
-        int BonusToDefense { get; }
-        int BonusToAttack { get; }
-        int BonusToIntelligence { get; }
+        byte BonusToHealth { get; }
+        byte BonusToDefense { get; }
+        byte BonusToAttack { get; }
+        byte BonusToIntelligence { get; }
         ItemClass ItemClass { get; }
         string ItemName { get; }
     }
